Перенёс созданиеАpi в index.js, Поменял метод возвратака ошибок в Api на _getResponseData, Изменил параметр в Section.renderItems
