+++ src/utilits/constant.js
@@ -7,6 +7,13 @@ const inputNote = popupFormInfo.querySelector('.popup__field_belong_note');
 const popupAvatar = document.querySelector('.popup_form_avatar');
 const buttonImage = document.querySelector('.profile__image-container');
 const buttonAdd = document.querySelector('.profile__add-button');
+const config = {
+	baseUrl: 'https://nomoreparties.co/v1/plus-cohort-23',
+	headers: {
+		authorization: '43b4a3de-4bbd-4be1-abd1-bf011e8ef6fa',
+			'Content-Type': 'application/json'
+	}
+}
 
 
 
@@ -14,4 +21,4 @@ const buttonAdd = document.querySelector('.profile__add-button');
 
 
 
-export { cardsContainer, popupFormCreate, buttonAdd, buttonInfo, popupFormInfo, inputName, inputNote, popupAvatar, buttonImage }
+export { cardsContainer, popupFormCreate, buttonAdd, buttonInfo, popupFormInfo, inputName, inputNote, popupAvatar, buttonImage, config }
